+++ src/Ampersand/FSpec/Motivations.hs
@@ -10,7 +10,7 @@ import Text.Pandoc
 -- The general idea is that an Ampersand declaration such as:
 --     PURPOSE RELATION r[A*B] IN ENGLISH
 --     {+This text explains why r[A*B] exists-}
--- produces the exact right text in the functional specification
+-- produces the exact right text in the functional design document.
 
 -- The class Motivated exists so that we can write the Haskell expression 'purposeOf fSpec l x'
 -- anywhere we like for every type of x that could possibly be motivated in an Purpose.