+++ ReleaseNotes.md
@@ -3,6 +3,7 @@
   
 ## Unreleased changes
   * [Issue #600](https://github.com/AmpersandTarski/Ampersand/issues/600) When a configuration file exists (.yaml) with the same name as the file being compiled, it will be used if no config file is explicitly mentioned in the command. 
+  * Rename "functional specification" to "functional design" for more realistic expectations with users.
 
 ## v3.8.0 (20 december 2016) 
   * [Issue #587](https://github.com/AmpersandTarski/Ampersand/issues/587) There is no need any longer to explicitly specify the language in a script. If omitted, IN ENGLISH will be used as default. 
@@ -19,7 +20,7 @@
 
 ## v3.7.2 (28 october 2016)
   * Some updates in the meatgrinder. (still experimental)
-  * Several changes in the generation of the functional specification document. (Less LaTex specific)
+  * Several changes in the generation of the functional design document. (Less LaTex specific)
 
 ## v3.7.1 (30 september 2016)
   * Upgrade version of Haskell compiler to 8.0.1 (automatically used via Stack)